Move agent_addr module from cli/ to api/ to fix layer violation
Problem: api/exec.py, api/events.py, and mngr_wait/api.py imported from
cli/agent_addr.py, violating the layered import convention (cli should depend
on api, not the other way around).
Fix: Move agent_addr.py and its tests from cli/ to api/, since it contains
domain logic (address parsing, host filtering) that belongs in the API layer.
All import references across the codebase are updated.
