+++ libs/mngr/imbue/mngr/api/agent_addr_test.py
@@ -6,13 +6,13 @@ import pluggy
 import pytest
 from click.testing import CliRunner
 
+from imbue.mngr.api.agent_addr import AgentAddress
+from imbue.mngr.api.agent_addr import _address_matches_agent_match
+from imbue.mngr.api.agent_addr import _address_matches_host
+from imbue.mngr.api.agent_addr import _post_filter_matches_by_addresses
+from imbue.mngr.api.agent_addr import filter_agents_by_host_constraint
+from imbue.mngr.api.agent_addr import parse_identifier_as_address
 from imbue.mngr.api.find import AgentMatch
-from imbue.mngr.cli.agent_addr import AgentAddress
-from imbue.mngr.cli.agent_addr import _address_matches_agent_match
-from imbue.mngr.cli.agent_addr import _address_matches_host
-from imbue.mngr.cli.agent_addr import _post_filter_matches_by_addresses
-from imbue.mngr.cli.agent_addr import filter_agents_by_host_constraint
-from imbue.mngr.cli.agent_addr import parse_identifier_as_address
 from imbue.mngr.cli.stop import stop
 from imbue.mngr.errors import AgentNotFoundError
 from imbue.mngr.primitives import AgentId
