+++ libs/mngr/imbue/mngr/api/exec.py
@@ -9,14 +9,14 @@ from imbue.imbue_common.frozen_model import FrozenModel
 from imbue.imbue_common.logging import log_call
 from imbue.imbue_common.logging import log_span
 from imbue.imbue_common.mutable_model import MutableModel
+from imbue.mngr.api.agent_addr import find_agent_by_address
+from imbue.mngr.api.agent_addr import find_agents_by_addresses
+from imbue.mngr.api.agent_addr import parse_identifier_as_address
 from imbue.mngr.api.discover import discover_hosts_and_agents
 from imbue.mngr.api.find import AgentMatch
 from imbue.mngr.api.find import ensure_host_started
 from imbue.mngr.api.find import group_agents_by_host
 from imbue.mngr.api.providers import get_provider_instance
-from imbue.mngr.cli.agent_addr import find_agent_by_address
-from imbue.mngr.cli.agent_addr import find_agents_by_addresses
-from imbue.mngr.cli.agent_addr import parse_identifier_as_address
 from imbue.mngr.config.data_types import MngrContext
 from imbue.mngr.errors import MngrError
 from imbue.mngr.errors import UserInputError
